# docs/_static/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/items/": {"get": {"summary": "Read Items", "operationId": "read_items_items__get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}, {"name": "q", "in": "query", "required": false, "schema": {"type": "string", "default": "", "title": "Q"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ItemBase"}, "title": "Response Read Items Items  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"summary": "Create Item", "description": "Insert a new item.", "operationId": "create_item_items__post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScanDataCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ItemBase"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/items/{item_id}": {"get": {"summary": "Read Item", "operationId": "read_item_items__item_id__get", "parameters": [{"name": "item_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Item Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ItemBase"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/tst/": {"post": {"summary": "Create Tst", "description": "Insert a TimeStampToken.", "operationId": "create_tst_tst__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TimeStampTokenCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TimeStampTokenCreate"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/tsts/": {"get": {"summary": "Read Tsts", "operationId": "read_tsts_tsts__get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/TimeStampToken"}, "title": "Response Read Tsts Tsts  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/tsts/{scan_uuid}": {"get": {"summary": "Get Tst", "operationId": "get_tst_tsts__scan_uuid__get", "parameters": [{"name": "scan_uuid", "in": "path", "required": true, "schema": {"title": "Scan Uuid"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TimeStampToken"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/system/stats/": {"get": {"summary": "Stats", "description": "Provides stats about the database.", "operationId": "stats_system_stats__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/system/info/": {"get": {"summary": "System Info", "description": "Provides information about the instance.", "operationId": "system_info_system_info__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ItemBase": {"properties": {"scan_data": {"$ref": "#/components/schemas/ScanData"}}, "type": "object", "required": ["scan_data"], "title": "ItemBase"}, "Meta": {"properties": {"uuid": {"type": "string", "title": "Uuid"}, "ts": {"type": "integer", "title": "Ts"}, "type": {"type": "string", "title": "Type"}}, "type": "object", "required": ["uuid", "ts", "type"], "title": "Meta"}, "Payload": {"properties": {"row": {"type": "string", "title": "Row"}}, "type": "object", "required": ["row"], "title": "Payload"}, "ScanData": {"properties": {"version": {"type": "string", "title": "Version"}, "format": {"type": "string", "title": "Format"}, "meta": {"$ref": "#/components/schemas/Meta"}, "payload": {"$ref": "#/components/schemas/Payload"}}, "type": "object", "required": ["version", "format", "meta", "payload"], "title": "ScanData"}, "ScanDataCreate": {"properties": {"version": {"type": "string", "title": "Version"}, "format": {"type": "string", "title": "Format"}, "meta": {"$ref": "#/components/schemas/Meta"}, "payload": {"$ref": "#/components/schemas/Payload"}}, "type": "object", "required": ["version", "format", "meta", "payload"], "title": "ScanDataCreate"}, "TimeStampToken": {"properties": {"scan_uuid": {"type": "string", "format": "uuid", "title": "Scan Uuid"}, "tst": {"type": "string", "format": "binary", "title": "Tst"}}, "type": "object", "required": ["scan_uuid", "tst"], "title": "TimeStampToken"}, "TimeStampTokenCreate": {"properties": {"scan_uuid": {"type": "string", "format": "uuid", "title": "Scan Uuid"}, "tst": {"type": "string", "format": "binary", "title": "Tst"}}, "type": "object", "required": ["scan_uuid", "tst"], "title": "TimeStampTokenCreate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}